+++ Sources/OsmEdit/OsmData/OAWay.h
@@ -18,12 +18,12 @@ NS_ASSUME_NONNULL_BEGIN
 @interface OAWay : OAEntity <OAEntityProtocol>
 
 -(id)initWithWay:(OAWay *)way;
--(id)initWithId:(long)identifier nodes:(NSArray<OANode *> *)nodes;
--(id)initWithId:(long)identifier latitude:(double)lat longitude:(double)lon ids:(NSArray<NSNumber *> *)nodeIds;
+-(id)initWithId:(long long)identifier nodes:(NSArray<OANode *> *)nodes;
+-(id)initWithId:(long long)identifier latitude:(double)lat longitude:(double)lon ids:(NSArray<NSNumber *> *)nodeIds;
 
--(void)addNodeById:(long)identifier;
--(long) getFirstNodeId;
--(long)getLastNodeId;
+-(void)addNodeById:(long long)identifier;
+-(long long) getFirstNodeId;
+-(long long)getLastNodeId;
 
 -(OANode *) getFirstNode;
 -(OANode *) getLastNode;
